refactor large InstancesViewController into separate ViewController per each Detail view this is makes individual web pages per each detail view, in order to improve load performance
