+++ src/main/resources/templates/instance-detail-view.html
@@ -92,52 +92,68 @@
 
         <nav>
             <div class="nav nav-tabs" id="nav-tab" role="tablist">
-                <a class="nav-item nav-link active" id="nav-home-tab" data-toggle="tab" href="#nav-variables" role="tab"
+                <a class="nav-item nav-link {{#content-variable-list-view}}active{{/content-variable-list-view}}"
+                   id="nav-home-tab" href="/views/instances/{{instance.processInstanceKey}}/variable-list" role="tab"
                    aria-controls="nav-home" aria-selected="true">Variables</a>
-                <a class="nav-item nav-link" id="nav-audit-tab" data-toggle="tab" href="#nav-audit-log" role="tab"
+                <a class="nav-item nav-link {{#content-audit-log-view}}active{{/content-audit-log-view}}"
+                   id="nav-audit-tab" href="/views/instances/{{instance.processInstanceKey}}/audit-log" role="tab"
                    aria-controls="nav-contact" aria-selected="false">Audit Log</a>
-                <a class="nav-item nav-link" id="nav-incidents-tab" data-toggle="tab" href="#nav-incidents" role="tab"
+                <a class="nav-item nav-link {{#content-incident-list-view}}active{{/content-incident-list-view}}"
+                   id="nav-incidents-tab" href="/views/instances/{{instance.processInstanceKey}}/incident-list" role="tab"
                    aria-controls="nav-profile" aria-selected="false">Incidents</a>
-                <a class="nav-item nav-link" id="nav-jobs-tab" data-toggle="tab" href="#nav-jobs" role="tab"
+                <a class="nav-item nav-link {{#content-job-list-view}}active{{/content-job-list-view}}"
+                   id="nav-jobs-tab" href="/views/instances/{{instance.processInstanceKey}}/job-list" role="tab"
                    aria-controls="nav-contact" aria-selected="false">Jobs</a>
-                <a class="nav-item nav-link" id="nav-message-subscriptions-tab" data-toggle="tab"
-                   href="#nav-message-subscriptions" role="tab" aria-controls="nav-contact" aria-selected="false">Message
+                <a class="nav-item nav-link {{#content-message-subscription-list-view}}active{{/content-message-subscription-list-view}}"
+                   id="nav-message-subscriptions-tab"
+                   href="/views/instances/{{instance.processInstanceKey}}/message-subscription-list" role="tab" aria-controls="nav-contact" aria-selected="false">Message
                     Subscriptions</a>
-                <a class="nav-item nav-link" id="nav-timers-tab" data-toggle="tab" href="#nav-timers" role="tab"
+                <a class="nav-item nav-link {{#content-timer-list-view}}active{{/content-timer-list-view}}"
+                   id="nav-timers-tab" href="/views/instances/{{instance.processInstanceKey}}/timer-list" role="tab"
                    aria-controls="nav-contact" aria-selected="false">Timers</a>
-                <a class="nav-item nav-link" id="nav-called-processes-tab" data-toggle="tab"
-                   href="#nav-called-processes" role="tab"
+                <a class="nav-item nav-link {{#content-called-processes-list-view}}active{{/content-called-processes-list-view}}"
+                   id="nav-called-processes-tab"
+                   href="/views/instances/{{instance.processInstanceKey}}/called-processes-list" role="tab"
                    aria-controls="nav-contact" aria-selected="false">Called Process Instances</a>
-                <a class="nav-item nav-link" id="nav-errors-tab" data-toggle="tab"
-                   href="#nav-errors" role="tab"
+                <a class="nav-item nav-link {{#content-error-list-view}}active{{/content-error-list-view}}"
+                   id="nav-errors-tab"
+                   href="/views/instances/{{instance.processInstanceKey}}/error-list" role="tab"
                    aria-controls="nav-contact" aria-selected="false">Errors</a>
             </div>
         </nav>
         <div class="tab-content detail-view-tab-content" id="nav-tabContent">
-            <div class="tab-pane fade show active" id="nav-variables" role="tabpanel" aria-labelledby="nav-home-tab">
+
+            {{#content-variable-list-view}}
                 {{>variable-list}}
-            </div>
-            <div class="tab-pane fade" id="nav-audit-log" role="tabpanel" aria-labelledby="nav-contact-tab">
+            {{/content-variable-list-view}}
+
+            {{#content-audit-log-view}}
                 {{>audit-log}}
-            </div>
-            <div class="tab-pane fade" id="nav-incidents" role="tabpanel" aria-labelledby="nav-profile-tab">
+            {{/content-audit-log-view}}
+
+            {{#content-incident-list-view}}
                 {{>incident-list}}
-            </div>
-            <div class="tab-pane fade" id="nav-jobs" role="tabpanel" aria-labelledby="nav-contact-tab">
+            {{/content-incident-list-view}}
+
+            {{#content-job-list-view}}
                 {{>job-list}}
-            </div>
-            <div class="tab-pane fade" id="nav-message-subscriptions" role="tabpanel" aria-labelledby="nav-contact-tab">
+            {{/content-job-list-view}}
+
+            {{#content-message-subscription-list-view}}
                 {{>message-subscription-list}}
-            </div>
-            <div class="tab-pane fade" id="nav-timers" role="tabpanel" aria-labelledby="nav-contact-tab">
+            {{/content-message-subscription-list-view}}
+
+            {{#content-timer-list-view}}
                 {{>timer-list}}
-            </div>
-            <div class="tab-pane fade" id="nav-called-processes" role="tabpanel" aria-labelledby="nav-contact-tab">
+            {{/content-timer-list-view}}
+
+            {{#content-called-processes-list-view}}
                 {{>called-processes-list}}
-            </div>
-            <div class="tab-pane fade" id="nav-errors" role="tabpanel" aria-labelledby="nav-contact-tab">
+            {{/content-called-processes-list-view}}
+
+            {{#content-error-list-view}}
                 {{>error-list}}
-            </div>
+            {{/content-error-list-view}}
 
         </div>
